Extract mock module.
